requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.1
dropbox==9.3.0
gunicorn==19.6.0
idna==2.8
img2pdf==0.3.3
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
qrcode==6.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
whitenoise==4.1.2
